pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inbound-inquiry-analyzer"
version = "0.1.0"
description = "Classify inbound inquiries and output a formatted XLSX workbook"
requires-python = ">=3.10"
dependencies = [
    "openpyxl>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pre-commit>=3.0",
]

[project.scripts]
inbound-inquiry-analyzer = "inbound_inquiry_analyzer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
